Changes from the last month or so, extremely lazy but it needs to be done

+++ src/main/java/tragicneko/tragicmc/entity/mob/EntityTox.java
@@ -326,7 +326,7 @@ public class EntityTox extends TragicMob {
 	public IEntityLivingData onSpawnWithEgg(IEntityLivingData data)
 	{
 		BiomeGenBase biome = this.worldObj.getBiomeGenForCoords((int) this.posX, (int) this.posZ);
-		this.setToxType(biome instanceof BiomeGenPaintedForest ? (byte) 1 : 0);
+		this.setToxType(biome instanceof BiomeGenPaintedForest ? (byte) 1 : (byte) 0);
 		return super.onSpawnWithEgg(data);
 	}
